Adds option for per client custom login page
The login script checks the client folder for a folder that matches the app name for the client being accessed (based on URL). If the folder is found, it uses the login.php file in that folder. Otherwise, it uses the default view from the login.php in the root folder of the application.